Moving subscriptions and addons to subscriptions folder, moving user subscriptions and toybox to users folder
