+++ src/Resources/Subscriptions/SubscriptionAddons/SubscriptionAddons.ts
@@ -1,4 +1,4 @@
-import AbstractResource from '../AbstractResource';
+import AbstractResource from '../../AbstractResource';
 
 export default class SubscriptionAddons extends AbstractResource {
     initialise(): void {
