+++ src/__tests__/Client.spec.ts
@@ -2,18 +2,18 @@ import { httpClient } from '../../test/mocks/HttpClient';
 import Client from '../Client';
 import Auth from '../Resources/Auth/Auth';
 import Users from '../Resources/Users/Users';
-import UserSubscriptions from '../Resources/UserSubscriptions/UserSubscriptions';
 import WACCOrders from '../Resources/Wacc/Orders/Orders';
 import WACCReturnOrders from '../Resources/Wacc/ReturnOrders/ReturnOrders';
 import WACCSearch from '../Resources/Wacc/Search/Search';
 import WACCUsers from '../Resources/Wacc/Users/Users';
 import Subscriptions from '../Resources/Subscriptions/Subscriptions';
-import SubscriptionAddons from '../Resources/SubscriptionAddons/SubscriptionAddons';
+import SubscriptionAddons from '../Resources/Subscriptions/SubscriptionAddons/SubscriptionAddons';
 import Codes from '../Resources/Codes/Codes';
-import Toybox from '../Resources/Toybox/Toybox';
+import Toybox from '../Resources/User/Toybox/Toybox';
 import Products from '../Resources/Products/Products';
 import GuestBasket from '../Resources/Guest/Basket/Basket';
 import UserBasket from '../Resources/User/Basket/Basket';
+import UserSubscriptions from '../Resources/User/Subscriptions/Subscriptions';
 
 describe('Client.ts', () => {
     let client: Client;
@@ -34,11 +34,9 @@ describe('Client.ts', () => {
         // @todo We could probably turn this in to a loop to make it easier
         expect(client.auth).toBeInstanceOf(Auth);
         expect(client.users).toBeInstanceOf(Users);
-        expect(client.userSubscriptions).toBeInstanceOf(UserSubscriptions);
-        expect(client.subscriptions).toBeInstanceOf(Subscriptions);
-        expect(client.subscriptionAddons).toBeInstanceOf(SubscriptionAddons);
+        expect(client.subscriptions.subscriptions).toBeInstanceOf(Subscriptions);
+        expect(client.subscriptions.addons).toBeInstanceOf(SubscriptionAddons);
         expect(client.codes).toBeInstanceOf(Codes);
-        expect(client.toybox).toBeInstanceOf(Toybox);
         expect(client.products).toBeInstanceOf(Products);
         expect(client.wacc.orders).toBeInstanceOf(WACCOrders);
         expect(client.wacc.returnOrders).toBeInstanceOf(WACCReturnOrders);
@@ -46,6 +44,8 @@ describe('Client.ts', () => {
         expect(client.wacc.users).toBeInstanceOf(WACCUsers);
         expect(client.guest.basket).toBeInstanceOf(GuestBasket);
         expect(client.user.basket).toBeInstanceOf(UserBasket);
+        expect(client.user.subscriptions).toBeInstanceOf(UserSubscriptions);
+        expect(client.user.toybox).toBeInstanceOf(Toybox);
     });
 
     it('can return the currently set feature options', () => {
